+++ src/model/ChatExtension.ts
@@ -14,6 +14,7 @@ As the AI assistant of this extension, you have these responsabilities:
 -You are provided with the user's commands history, and you can convert it, or parts of it, in snakemake or make rules, following user's requests.
 -You also are provided with some information of the current state of the extension software and you use it to help the user understand the extension's behavior.
 -Whether you write rules in Snakemake or in Make depends on your current setting.
+-In Snakemake, a best practice requires each rule to have a log directive. By default the extension do that, but user can disable it by changing the settings related to Snakemake best practices.
 -You can also return these commands to help the user:
     [Start listening to bash commands](command:start-listening)   #Start listening to bash commands
     [Stop listening to bash commands](command:stop-listening)   #Stop listening to bash commands
@@ -26,6 +27,8 @@ As the AI assistant of this extension, you have these responsabilities:
     [Open settings](command:workbench.action.openSettings?"snakemaker.allowLogging")   #Open the settings to enable/disable logging
     [Open settings](command:workbench.action.openSettings?"snakemaker.rulesOutputFormat")   #Open the settings to switch between Snakemake and Make rules
     [Open settings](command:workbench.action.openSettings?"snakemaker.keepHistoryBetweenSessions")   #Open the settings to enable-disable keeping history between sessions
+    [Open settings](command:workbench.action.openSettings?"snakemaker.snakemakeBestPractices")   #Open the settings related to how the snakemake rules are written, if formalisms related to best practices are followed.
+
 The command history-set?NEW_HISTORY_JSON sets a new history. Use it if the user asks to perform changes. You have to: 1- Briefly tell the user which changes you are performing (DO NOT show the entire JSON with the new history, explain briefly which things you're modifying). 2-Valorize NEW_HISTORY_JSON as the modified version of the history you are provided as HISTORY OF RECORDED BASH COMMANDS. You can also use this example as a template of how the history is organized:EXAMPLE OF HISTORY, with one unimportant command, one important command and one composite, important command: {"history":[{"commands":[{"command":"dir","exitStatus":0,"output":"-","inputs":"-","important":false,"index":2,"temporary":false,"rule_name":"list_directory"}],"index":3,"rule_name":""},{"commands":[{"command":"catinput.txt|wc-l>output.txt","exitStatus":0,"output":"\"output.txt\"","inputs":"\"input.txt\"","important":true,"index":15,"temporary":false,"rule_name":"count_lines"}],"index":16,"rule_name":""},{"commands":[{"command":"mkdirresults","exitStatus":0,"output":"results","inputs":"-","important":true,"index":10,"temporary":false,"rule_name":"create_results_directory"},{"command":"catinput.txt|wc-l>results/output.txt","exitStatus":0,"output":"\"results/output.txt\"","inputs":"\"input.txt\"","important":true,"index":13,"temporary":false,"rule_name":"\"count_lines\""}],"index":9,"rule_name":"make_results_and_outputs"}]}
 Please note the command history-set can be used only through the chat, not manually from command palette.
 If the user asks you to do something not doable with these commands, tell him you can't do it yourself and explain how he can do it himself.
@@ -104,6 +107,7 @@ HERE IS THE HISTORY:`;
     stream: vscode.ChatResponseStream,token: vscode.CancellationToken){
         const rule_format = vscode.workspace.getConfiguration('snakemaker').get('rulesOutputFormat', "Snakemake");
         const mustStash = vscode.workspace.getConfiguration('snakemaker').get('keepHistoryBetweenSessions', false);
+        const containsLogField = vscode.workspace.getConfiguration('snakemaker').get('snakemakeBestPracticesSetLogFieldInSnakemakeRules', false);
         const messages = [
             vscode.LanguageModelChatMessage.User(ChatExtension.BASE_PROMPT),
             vscode.LanguageModelChatMessage.User(ChatExtension.BASE_PROMPT_EXTENSION_USAGE),
@@ -111,7 +115,7 @@ HERE IS THE HISTORY:`;
                 ChatExtension.BASH_HISTORY_INTRODUCTION + this.history.history_for_the_chat()
             ),
             vscode.LanguageModelChatMessage.User(
-                `Additional extension info: currently listening to bash commands: ${this.viewModel.isListening}. Copilot active: ${this.viewModel.isCopilotActive()}  Currently changing model: ${this.viewModel.isChangingModel}. Models available: ${this.viewModel.llm.models.map((m) => m.get_name())}. Active model: ${this.viewModel.llm.models[this.viewModel.llm.current_model]?.get_name()||'none'} - Logging status: ${SnkmakerLogger.logger_status()} - Current rule format: ${rule_format} - Keep history between sessions: ${mustStash}`
+                `Additional extension info: currently listening to bash commands: ${this.viewModel.isListening}. Copilot active: ${this.viewModel.isCopilotActive()}  Currently changing model: ${this.viewModel.isChangingModel}. Models available: ${this.viewModel.llm.models.map((m) => m.get_name())}. Active model: ${this.viewModel.llm.models[this.viewModel.llm.current_model]?.get_name()||'none'} - Logging status: ${SnkmakerLogger.logger_status()} - Current rule format: ${rule_format} - Snakemake rules contains Log directive: ${containsLogField} - Keep history between sessions: ${mustStash}`
             )
         ];
         // get the previous messages