+++ CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Improved prompts - now the model prefers to output generic rules using wildcards.
+- Improved Snakemake rule generation regarding the best practices: now rules contains Log directive. 
+- Log directive in Snakemake rule can be disabled in settings.
 
 
 ## [0.1.2] - 2025-01-07